Factored out InstanceDatabaseMigrator to follow the same structre of classes we have in acase of forms db
