+++ collect_app/src/main/java/org/odk/collect/android/database/InstancesDatabaseHelper.java
@@ -19,176 +19,41 @@ package org.odk.collect.android.database;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import org.odk.collect.android.instances.Instance;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
-import org.odk.collect.android.utilities.SQLiteUtils;
-
-import java.util.Arrays;
-import java.util.List;
 
 import timber.log.Timber;
-
-import static android.provider.BaseColumns._ID;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.CAN_EDIT_WHEN_COMPLETE;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.DELETED_DATE;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.DISPLAY_NAME;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.GEOMETRY;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.GEOMETRY_TYPE;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.INSTANCE_FILE_PATH;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.JR_FORM_ID;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.JR_VERSION;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.LAST_STATUS_CHANGE_DATE;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.STATUS;
-import static org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns.SUBMISSION_URI;
 
 /**
  * This class helps open, create, and upgrade the database file.
  */
 public class InstancesDatabaseHelper extends SQLiteOpenHelper {
     private static final String DATABASE_NAME = "instances.db";
-    public static final String INSTANCES_TABLE_NAME = "instances";
-
     public static final int DATABASE_VERSION = 6;
 
-    private static final String[] COLUMN_NAMES_V5 = {_ID, DISPLAY_NAME, SUBMISSION_URI, CAN_EDIT_WHEN_COMPLETE,
-            INSTANCE_FILE_PATH, JR_FORM_ID, JR_VERSION, STATUS, LAST_STATUS_CHANGE_DATE, DELETED_DATE};
-
-    private static final String[] COLUMN_NAMES_V6 = {_ID, DISPLAY_NAME, SUBMISSION_URI,
-        CAN_EDIT_WHEN_COMPLETE, INSTANCE_FILE_PATH, JR_FORM_ID, JR_VERSION, STATUS,
-        LAST_STATUS_CHANGE_DATE, DELETED_DATE, GEOMETRY, GEOMETRY_TYPE};
-
-    public static final String[] CURRENT_VERSION_COLUMN_NAMES = COLUMN_NAMES_V6;
+    private final InstanceDatabaseMigrator instanceDatabaseMigrator;
 
     public InstancesDatabaseHelper() {
         super(new DatabaseContext(new StoragePathProvider().getDirPath(StorageSubdirectory.METADATA)), DATABASE_NAME, null, DATABASE_VERSION);
+        instanceDatabaseMigrator = new InstanceDatabaseMigrator();
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        createInstancesTableV5(db, INSTANCES_TABLE_NAME);
-        upgradeToVersion6(db, INSTANCES_TABLE_NAME);
+        instanceDatabaseMigrator.onCreate(db);
     }
 
-    /**
-     * Upgrades the database.
-     *
-     * When a new migration is added, a corresponding test case should be added to
-     * InstancesDatabaseHelperTest by copying a real database into assets.
-     */
-    @SuppressWarnings({"checkstyle:FallThrough"})
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         Timber.i("Upgrading database from version %d to %d", oldVersion, newVersion);
-
-        switch (oldVersion) {
-            case 1:
-                upgradeToVersion2(db);
-            case 2:
-                upgradeToVersion3(db);
-            case 3:
-                upgradeToVersion4(db);
-            case 4:
-                upgradeToVersion5(db);
-            case 5:
-                upgradeToVersion6(db, INSTANCES_TABLE_NAME);
-                break;
-            default:
-                Timber.i("Unknown version %d", oldVersion);
-        }
-
+        instanceDatabaseMigrator.onUpgrade(db, oldVersion, newVersion);
         Timber.i("Upgrading database from version %d to %d completed with success.", oldVersion, newVersion);
     }
 
     @Override
     public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         Timber.i("Downgrading database from version %d to %d", oldVersion, newVersion);
-
-        String temporaryTableName = INSTANCES_TABLE_NAME + "_tmp";
-        createInstancesTableV5(db, temporaryTableName);
-        upgradeToVersion6(db, temporaryTableName);
-
-        dropObsoleteColumns(db, CURRENT_VERSION_COLUMN_NAMES, temporaryTableName);
+        instanceDatabaseMigrator.onDowngrade(db, oldVersion, newVersion);
         Timber.i("Downgrading database from version %d to %d completed with success.", oldVersion, newVersion);
-    }
-
-    private void upgradeToVersion2(SQLiteDatabase db) {
-        if (!SQLiteUtils.doesColumnExist(db, INSTANCES_TABLE_NAME, CAN_EDIT_WHEN_COMPLETE)) {
-            SQLiteUtils.addColumn(db, INSTANCES_TABLE_NAME, CAN_EDIT_WHEN_COMPLETE, "text");
-
-            db.execSQL("UPDATE " + INSTANCES_TABLE_NAME + " SET "
-                    + CAN_EDIT_WHEN_COMPLETE + " = '" + true
-                    + "' WHERE " + STATUS + " IS NOT NULL AND "
-                    + STATUS + " != '" + Instance.STATUS_INCOMPLETE
-                    + "'");
-        }
-    }
-
-    private void upgradeToVersion3(SQLiteDatabase db) {
-        SQLiteUtils.addColumn(db, INSTANCES_TABLE_NAME, JR_VERSION, "text");
-    }
-
-    private void upgradeToVersion4(SQLiteDatabase db) {
-        SQLiteUtils.addColumn(db, INSTANCES_TABLE_NAME, DELETED_DATE, "date");
-    }
-
-    /**
-     * Upgrade to version 5. Prior versions of the instances table included a {@code displaySubtext}
-     * column which was redundant with the {@link org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns#STATUS} and
-     * {@link org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns#LAST_STATUS_CHANGE_DATE} columns and included
-     * unlocalized text. Version 5 removes this column.
-     */
-    private void upgradeToVersion5(SQLiteDatabase db) {
-        String temporaryTableName = INSTANCES_TABLE_NAME + "_tmp";
-
-        // onDowngrade in Collect v1.22 always failed to clean up the temporary table so remove it now.
-        // Going from v1.23 to v1.22 and back to v1.23 will result in instance status information
-        // being lost.
-        SQLiteUtils.dropTable(db, temporaryTableName);
-
-        createInstancesTableV5(db, temporaryTableName);
-        dropObsoleteColumns(db, COLUMN_NAMES_V5, temporaryTableName);
-    }
-
-    /**
-     * Use the existing temporary table with the provided name to only keep the given relevant
-     * columns, dropping all others.
-     *
-     * NOTE: the temporary table with the name provided is dropped.
-     *
-     * The move and copy strategy is used to overcome the fact that SQLITE does not directly support
-     * removing a column. See https://sqlite.org/lang_altertable.html
-     *
-     * @param db                    the database to operate on
-     * @param relevantColumns       the columns relevant to the current version
-     * @param temporaryTableName    the name of the temporary table to use and then drop
-     */
-    private void dropObsoleteColumns(SQLiteDatabase db, String[] relevantColumns, String temporaryTableName) {
-        List<String> columns = SQLiteUtils.getColumnNames(db, INSTANCES_TABLE_NAME);
-        columns.retainAll(Arrays.asList(relevantColumns));
-        String[] columnsToKeep = columns.toArray(new String[0]);
-
-        SQLiteUtils.copyRows(db, INSTANCES_TABLE_NAME, columnsToKeep, temporaryTableName);
-        SQLiteUtils.dropTable(db, INSTANCES_TABLE_NAME);
-        SQLiteUtils.renameTable(db, temporaryTableName, INSTANCES_TABLE_NAME);
-    }
-
-    private void upgradeToVersion6(SQLiteDatabase db, String name) {
-        SQLiteUtils.addColumn(db, name, GEOMETRY, "text");
-        SQLiteUtils.addColumn(db, name, GEOMETRY_TYPE, "text");
-    }
-
-    private void createInstancesTableV5(SQLiteDatabase db, String name) {
-        db.execSQL("CREATE TABLE IF NOT EXISTS " + name + " ("
-                + _ID + " integer primary key, "
-                + DISPLAY_NAME + " text not null, "
-                + SUBMISSION_URI + " text, "
-                + CAN_EDIT_WHEN_COMPLETE + " text, "
-                + INSTANCE_FILE_PATH + " text not null, "
-                + JR_FORM_ID + " text not null, "
-                + JR_VERSION + " text, "
-                + STATUS + " text not null, "
-                + LAST_STATUS_CHANGE_DATE + " date not null, "
-                + DELETED_DATE + " date );");
     }
 }
